feat(Modal): Separate section of MessageDialog into Modal component (#126)
